+++ examples/tutorial_derive/03_01_flag_count.rs
@@ -4,7 +4,7 @@ use clap::Parser;
 #[clap(author, version, about, long_about = None)]
 struct Cli {
     #[clap(short, long, action = clap::ArgAction::Count)]
-    verbose: u64,
+    verbose: u8,
 }
 
 fn main() {
